Feat/account tables (#43)
This add:
- Tables to store account_updates transactions
- auxiliary public_keys table 
- method to retrieve historical updates for a specific account.
- endpoint to retrieve historical updates for accounts.

Even though we index data from **tx_inti_account**, which outlines the
initial state of an account, it notably lacks an account_id. This
omission is logical, considering the account is yet to be created.
Consequently, we can't establish a connection between an update and an
init_account transaction. For this reason, we've chosen to focus solely
on tracking account updates. To accommodate this, we've created a
specific endpoint that allows users to retrieve this information by
submitting an account_id (Address).

@Fraccaman more insights on this would be helpful, how could we link a
tx_init_account to a later tx_update_account?