feat: added return code to database (#55)
This PR add a new feature that allow fetch return code for
decrypted/inner transactions. This info will be then added to the
database.

It required adding a call to the tendermint `/block_results` enpoint.
This specific endpoint return events that happened with the transaction.
We can find the `code` (meaning the return code) there.