fix: remove foreign key constraint  (#77)
fix issue #76 

Remove the foreign key constraint on the secondary tables that collect
the decoded transaction data. It should speed up things and remove the
fk violation error because the indexer try to insert the decoded
transaction before the raw transaction info.

We are assuming They would be there because they come from the same
source of information.