Do not set INDEXER_CONFIG_PATH by default in Dockerfile (#149)
This allows the docker image to be configured via env variables instead
of a config file. (When `INDEXER_CONFIG_PATH` is set, the application
will only use the file and will not attempt to read its config from env
vars.) This is useful for environments like Google Cloud Run where it is
simpler to set env vars than to mount a file into the container.