Propagades timeout error (#162)
The enhancement shifts the timeout mechanism from the block producer to
the consumer side. This change ensures that if the block producer
stalls, the consumer, instead of indefinitely waiting on an empty
channel, will encounter a timeout. This timeout triggers an error that
is then easily propagated back to the main function. Consequently, this
allows the entire indexing process to terminate gracefully, rather than
abruptly due to a panic.