fix: Update views tx_vote_proposal and tx_transfer (#185)
The views tx_vote_proposal and tx_transfer took in all transactions
regardless of status (failed/succeeded). Updated views to only use
successful transactions, as in former version tables. Note: An alternate
approach might be to make the views richer and include fail/success
status in tables. For now this is a quick fix.