feat: add code_type value and memo data for transactions (#197)
This PR add 2 new columns to the transactions table : `code_type` that
translate the code in a human readable string and `memo` which contains
the memo data associated with the transaction.

To avoid issue when deploying this new feature the table is altered if
needed. Only new indexed block data will have those new values saved. To
have previous data a fully resync is needed.